+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Change Log
 
 ## 0.1.1 Under development
 
+- Enh #34: Add `BunManager` class to manage the `Bun` instances (@terabytesoftw)
+
 ## 0.1.0 January 21, 2024
 
 - Initial release.